test(storage): make the backend harness honest
