+++ localdisk-storage/docs/test_plan.md
@@ -22,14 +22,14 @@ Trait-level contract testing (save, load, delete, recovery scan) is covered by t
 | :--- | :--- | :--- |
 | **7.1.1** | Configurable location for Local Disk bundle storage | LD-01 |
 | **7.1.2** | Configurable maximum total size for Local Disk storage | LD-05 |
-| **7.1** | Store/retrieve bundles | [`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) Suite D (BLOB-01..04) |
+| **7.1** | Store/retrieve bundles | [`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) Suite D (BLOB-01..05) |
 | **7.3** | Recovery after restart | [`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) BLOB-04 |
 
 ## 3. Generic Harness Coverage
 
 This backend is registered in the storage harness with `storage_blob_tests!(localdisk, ...)`. The following suite runs against localdisk:
 
-- Suite D: Payload Operations (BLOB-01..04)
+- Suite D: Payload Operations (BLOB-01..05)
 
 ## 4. Backend-Specific Test Cases
 
@@ -41,7 +41,7 @@ This backend is registered in the storage harness with `storage_blob_tests!(loca
 | **LD-02** | **Recovery cleanup** | `storage.rs` | 1. Create storage dir with valid bundles, `.tmp` files, zero-byte files, and empty subdirectories.<br>2. Call `recover()`. | Valid bundles returned; `.tmp` files, zero-byte files, and empty dirs cleaned up. |
 | **LD-03** | **Filesystem structure** | `storage.rs` | 1. Save multiple bundles.<br>2. Inspect directory tree. | Files distributed in `xx/yy/` two-level subdirectories; filename collisions resolved without error. |
 | **LD-04** | **Atomic save** | `storage.rs` | 1. Call `save(data)` with `fsync=true`.<br>2. Verify file written to `.tmp` then renamed. | No partial files visible; rename is atomic. |
-| **LD-05** | **Write failure handling** | `storage.rs` | 1. Make store directory read-only.<br>2. Call `save(data)`. | Graceful error returned, not panic. |
+| **LD-05** | **Write failure handling** | `storage.rs` | _Unix only._<br>1. Make store directory read-only.<br>2. Probe that writes are actually blocked (a privileged process bypasses permission bits; the assertion is skipped in that case).<br>3. Call `save(data)`. | Graceful error returned, not panic. |
 
 ## 5. Execution
