+++ postgres-storage/docs/test_plan.md
@@ -20,22 +20,22 @@ Trait-level contract testing (CRUD, polling, ordering, state transitions, recove
 
 | Ref | Description | Verified By |
 | :--- | :--- | :--- |
-| **8.1** | Store additional metadata in a remote PostgreSQL database instance | [`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) Suite A–C (META-01..17) |
-| **8.2** | Restart the system and recover state from a remote PostgreSQL instance | [`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) META-05, META-12, META-13 |
+| **8.1** | Store additional metadata in a remote PostgreSQL database instance | [`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) Suite A–C (META-01..11, META-13..17) |
+| **8.2** | Restart the system and recover state from a remote PostgreSQL instance | [`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) META-05, META-13 |
 
 ## 3. Generic Harness Coverage
 
-This backend is registered in the storage harness with `storage_meta_tests_async!(postgres, ...)` plus a dedicated `meta_05_confirm_exists` recovery test. The following suites run against PostgreSQL:
+This backend is registered in the storage harness with `storage_meta_tests!(postgres, ...)` plus the recovery suite `storage_meta_recovery_tests!(postgres_recovery, ...)`. The following suites run against PostgreSQL:
 
 - Suite A: Basic CRUD Operations (META-01..04, META-15, META-17)
 - Suite B: Polling & Ordering (META-06..10, META-14)
-- Suite C: State Transitions & Bulk Ops (META-05, META-11..13, META-16)
+- Suite C: State Transitions & Bulk Ops (META-05, META-11, META-13, META-16)
 
 Requires `--features postgres` and a running PostgreSQL instance (default: `postgresql://hardy:hardy@localhost:5432`).
 
 ## 4. Backend-Specific Test Rationale
 
-No unit tests are planned for this crate. The `MetadataStorage` trait contract is fully verified by the generic harness against a real PostgreSQL instance (17 tests, feature-gated behind `--features postgres`). The backend-specific scenarios listed below are inherently integration concerns — connection pooling, migration, concurrent access, and connection failure all depend on a live database and are effectively testing `sqlx` and PostgreSQL behaviour rather than Hardy code. Adding crate-level unit tests would duplicate the harness coverage or test third-party library semantics.
+No unit tests are planned for this crate. The `MetadataStorage` trait contract is fully verified by the generic harness against a real PostgreSQL instance (16 tests, feature-gated behind `--features postgres`). The backend-specific scenarios listed below are inherently integration concerns — connection pooling, migration, concurrent access, and connection failure all depend on a live database and are effectively testing `sqlx` and PostgreSQL behaviour rather than Hardy code. Adding crate-level unit tests would duplicate the harness coverage or test third-party library semantics.
 
 If backend-specific integration tests are needed in future (e.g. for schema upgrade verification across releases), they should be added to the storage harness as new feature-gated tests, reusing the existing per-test database isolation infrastructure.
