+++ s3-storage/docs/test_coverage_report.md
@@ -9,11 +9,11 @@
 
 ## 1. LLR Coverage Summary (Requirements Verification Matrix)
 
-The `BundleStorage` trait contract is verified by the shared storage harness (4 tests, all pass when an S3-compatible endpoint is available). Recovery requirements are satisfied by the harness test exercising `BundleStorage::recover()` (BLOB-04), with the BPA's `storage/recover.rs` orchestrating trait methods during restart.
+The `BundleStorage` trait contract is verified by the shared storage harness (5 tests, all pass when an S3-compatible endpoint is available). Recovery requirements are satisfied by the harness test exercising `BundleStorage::recover()` (BLOB-04), with the BPA's `storage/recover.rs` orchestrating trait methods during restart.
 
 | Part 4 Ref | Requirement | Result | Verified By |
 | :--- | :--- | :--- | :--- |
-| 9.1 | Bundle storage | **Pass** | BLOB-01..04 ([`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) §5) |
+| 9.1 | Bundle storage | **Pass** | BLOB-01..05 ([`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) §5) |
 | 9.1.1 | Configurable location/credentials | **Pass** | Harness setup ([`PLAN-S3-01`](test_plan.md) §5) — every run supplies the endpoint and credentials via configuration; CI runs against MinIO |
 | 9.1.2 | Configurable maximum total size | **Not tested** | Not planned |
 | 9.1.4 | Common S3 APIs | **Pass** | By design (`aws-sdk-s3`) |
@@ -23,9 +23,9 @@ The `BundleStorage` trait contract is verified by the shared storage harness (4 
 
 ### Generic harness tests (via `tests/storage/`)
 
-4 integration tests run against S3/MinIO via `storage_blob_tests_async!(s3, ...)`. See [`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) §3.5 for registration details and §5 for test scenarios.
+5 integration tests run against S3/MinIO via `storage_blob_tests!(s3, ...)`. See [`PLAN-STORE-01`](../../tests/storage/docs/test_plan.md) §3.5 for registration details and §5 for test scenarios.
 
-Each test uses a unique key prefix (`test-{uuid}`) for isolation within the shared bucket. All 4 pass. No failures or skips.
+Each test uses a unique key prefix (`test-{uuid}`) for isolation within the shared bucket, and a guard removes the prefix's objects (best-effort) when the test completes. All 5 pass. No failures or skips.
 
 ### Backend-specific tests
 
@@ -37,8 +37,8 @@ Coverage is measured against [`PLAN-S3-01`](test_plan.md) and [`PLAN-STORE-01`](
 
 | Source | Scope | Planned | Implemented | Status |
 | :--- | :--- | :--- | :--- | :--- |
-| `PLAN-STORE-01` Suite D (BLOB-01..04) | BundleStorage trait contract | 4 | 4 | Complete |
-| **Total** | | **4** | **4** | **100%** |
+| `PLAN-STORE-01` Suite D (BLOB-01..05) | BundleStorage trait contract | 5 | 5 | Complete |
+| **Total** | | **5** | **5** | **100%** |
 
 ## 4. Line Coverage
 
@@ -46,4 +46,4 @@ Line coverage is not measurable for this crate. All verification runs through th
 
 ## 5. Conclusion
 
-4 integration tests verify the full `BundleStorage` trait contract through the shared storage harness (100% of planned scenarios). All payload operations (save, load, delete, recovery scan) pass against a real S3-compatible endpoint with per-test prefix isolation. No backend-specific unit tests are planned — see [`PLAN-S3-01` §4](test_plan.md) for rationale.
+5 integration tests verify the full `BundleStorage` trait contract through the shared storage harness (100% of planned scenarios). All payload operations (save, load, delete, recovery scan, repeatable load) pass against a real S3-compatible endpoint with per-test prefix isolation. No backend-specific unit tests are planned — see [`PLAN-S3-01` §4](test_plan.md) for rationale.
