+++ s3-storage/src/lib.rs
@@ -133,41 +133,3 @@ impl<'de> serde::Deserialize<'de> for PartSize {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::PartSize;
-
-    #[test]
-    fn multipart_threshold_holds_the_s3_bound() {
-        use super::MultipartThreshold;
-
-        assert_eq!(
-            MultipartThreshold::new(0).map(MultipartThreshold::get),
-            Some(0)
-        );
-
-        #[cfg(target_pointer_width = "64")]
-        {
-            assert!(
-                MultipartThreshold::new((MultipartThreshold::MAX_BYTES as usize) + 1).is_none()
-            );
-            assert!(MultipartThreshold::new(MultipartThreshold::MAX_BYTES as usize).is_some());
-        }
-    }
-
-    #[test]
-    fn part_size_holds_the_s3_bounds() {
-        assert!(PartSize::new(5 * 1024 * 1024 - 1).is_none());
-        assert_eq!(
-            PartSize::new(5 * 1024 * 1024).map(PartSize::get),
-            Some(5 * 1024 * 1024)
-        );
-
-        #[cfg(target_pointer_width = "64")]
-        {
-            assert!(PartSize::new((PartSize::MAX_BYTES as usize) + 1).is_none());
-            assert!(PartSize::new(PartSize::MAX_BYTES as usize).is_some());
-        }
-    }
-}
